requirements.txt
django
djangorestframework
django-cors-headers
django-filter
iyzipay
Pillow
django-rest-knox